# openapi.yaml
openapi: 3.0.3
info:
  title: HRX API
  description: |
    API para gerenciamento de profissionais, fornecedores e projetos de eventos.

    ## Autenticação
    Todas as rotas requerem autenticação via Clerk. O token JWT deve ser enviado no header Authorization.

    ## Rate Limiting
    - Leitura: 100 requests/minuto
    - Escrita: 20 requests/minuto

    ## Otimizações
    - Índices geográficos para buscas por localização
    - RPC functions para queries complexas
    - Cache com Redis para dados frequentes
  version: 1.0.0
  contact:
    name: HRX Tecnologia
    email: contato@hrxeventos.com.br
  license:
    name: Proprietary
servers:
  - url: https://hrx.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT do Clerk
  schemas:
...
    EventProject:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_number:
          type: string
          example: "PRJ-2025-001"
        event_name:
          type: string
        event_type:
          type: string
        event_date:
          type: string
          format: date
        client_name:
          type: string
        venue_city:
          type: string
        venue_state:
          type: string
        status:
          type: string
          enum: [new, analyzing, quoting, quoted, proposed, approved, in_execution, completed, cancelled]
        is_urgent:
          type: boolean
        total_cost:
          type: number
          format: double
        total_client_price:
          type: number
          format: double
        professionals_needed:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              quantity:
                type: integer
        equipment_needed:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              quantity:
                type: integer
...
paths:
...
  /admin/event-projects:
    get:
      summary: Listar projetos de eventos
      description: |
        Lista projetos com filtros e paginação.

        **Otimização:** Select otimizado (apenas campos necessários)
      tags:
        - Event Projects
      security:
        - ClerkAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: is_urgent
          in: query
          schema:
            type: boolean
        - name: client_name
          in: query
          schema:
            type: string
        - name: event_type
          in: query
          schema:
            type: string
        - name: venue_city
          in: query
          schema:
            type: string
        - name: venue_state
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de projetos
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventProject'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      summary: Criar novo projeto
      tags:
        - Event Projects
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventProject'
      responses:
        '201':
          description: Projeto criado
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
...
security:
  - ClerkAuth: []